update: fix gcc warnings part 1

+++ sdk/lib/ucrt/filesystem/splitpath.cpp
@@ -313,10 +313,10 @@ static void __cdecl common_splitpath(
 {
     component_buffers<Character> components =
     {
-        drive,     drive     ? _MAX_DRIVE : 0,
-        directory, directory ? _MAX_DIR   : 0,
-        file_name, file_name ? _MAX_FNAME : 0,
-        extension, extension ? _MAX_EXT   : 0
+        drive,     drive     ? static_cast<size_t>(_MAX_DRIVE) : 0u,
+        directory, directory ? static_cast<size_t>(_MAX_DIR)   : 0u,
+        file_name, file_name ? static_cast<size_t>(_MAX_FNAME) : 0u,
+        extension, extension ? static_cast<size_t>(_MAX_EXT)   : 0u
     };
 
     common_splitpath_internal(path, &components, [](Character* const buffer, size_t const buffer_count)
